Refactor package structure and add ArrayList inspection
Moved classes from "dayeight.taskone" to "dayeight.task" for better package organization. Introduced a new class `ArrayListElementsDataChcking` to inspect internal fields of `ArrayList`. Enhanced `ArrayListForEachLoop` with additional element manipulation methods and simplified loop logic.
